Cambia la vista para que se trabaje con BigDecimal en lugar de double.

+++ src/main/java/sic/modelo/Cliente.java
@@ -1,6 +1,7 @@
 package sic.modelo;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -26,7 +27,7 @@ public class Cliente implements Serializable {
     private Usuario credencial;
     private boolean eliminado;
     private boolean predeterminado;
-    private Double saldoCuentaCorriente;
+    private BigDecimal saldoCuentaCorriente;
     private Date fechaUltimoMovimiento;
 
     @Override
